python 3.6, skip steps if done already

+++ scripts/download_data.sh
@@ -9,24 +9,34 @@ data=$base/data
 
 mkdir -p $data
 
-source $base/venvs/sockeye-autopilot/bin/activate
+for unused in pseudo_loop; do
 
-sockeye-autopilot --task wmt14_en_de --model none --workspace $data/sockeye_autopilot
+    if [[ -s $data/test.bpe.$src ]]; then
+      echo "File exists: $data/test.bpe.$src"
+      echo "Skipping. Delete files to repeat step."
+      continue
+    fi
 
-(cd $data/sockeye_autopilot/systems/wmt14_en_de/data/bpe && gunzip *)
+    source $base/venvs/sockeye-autopilot/bin/activate
 
-ln -snf $data/sockeye_autopilot/systems/wmt14_en_de/data/bpe/train.src $data/train.bpe.$src
-ln -snf $data/sockeye_autopilot/systems/wmt14_en_de/data/bpe/train.trg $data/train.bpe.$trg
+    sockeye-autopilot --task wmt14_en_de --model none --workspace $data/sockeye_autopilot
 
-ln -snf $data/sockeye_autopilot/systems/wmt14_en_de/data/bpe/dev.src $data/dev.bpe.$src
-ln -snf $data/sockeye_autopilot/systems/wmt14_en_de/data/bpe/dev.trg $data/dev.bpe.$trg
+    (cd $data/sockeye_autopilot/systems/wmt14_en_de/data/bpe && gunzip *)
 
-ln -snf $data/sockeye_autopilot/systems/wmt14_en_de/data/bpe/test.0.src $data/test.bpe.$src
-ln -snf $data/sockeye_autopilot/systems/wmt14_en_de/data/bpe/test.0.trg $data/test.bpe.$trg
+    ln -snf $data/sockeye_autopilot/systems/wmt14_en_de/data/bpe/train.src $data/train.bpe.$src
+    ln -snf $data/sockeye_autopilot/systems/wmt14_en_de/data/bpe/train.trg $data/train.bpe.$trg
 
-# sizes
-echo "Sizes of corpora:"
-wc -l $data/*
+    ln -snf $data/sockeye_autopilot/systems/wmt14_en_de/data/bpe/dev.src $data/dev.bpe.$src
+    ln -snf $data/sockeye_autopilot/systems/wmt14_en_de/data/bpe/dev.trg $data/dev.bpe.$trg
 
-# sanity checks
-echo "At this point, please make sure that 1) number of lines are as expected, 2) language suffixes are correct and 3) files are parallel"
+    ln -snf $data/sockeye_autopilot/systems/wmt14_en_de/data/bpe/test.0.src $data/test.bpe.$src
+    ln -snf $data/sockeye_autopilot/systems/wmt14_en_de/data/bpe/test.0.trg $data/test.bpe.$trg
+
+    # sizes
+    echo "Sizes of corpora:"
+    wc -l $data/*
+
+    # sanity checks
+    echo "At this point, please make sure that 1) number of lines are as expected, 2) language suffixes are correct and 3) files are parallel"
+
+done
